backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
ccxt==4.1.44
pandas==2.1.3
TA-Lib
APScheduler==3.10.4
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.25.2
numpy==1.25.2
python-dotenv==1.0.0
alembic==1.12.1
redis==5.0.1
celery==5.3.4
requests[socks]==2.31.0
pysocks==1.7.1
